// src/client/api.ts
const BASE_PATH = "http://localhost:5540".replace(/\/+$/, "");

/**
 *
 * @export
 */
export const COLLECTION_FORMATS = {
    csv: ",",
    ssv: " ",
    tsv: "\t",
    pipes: "|",
};

/**
 *
 * @export
 * @interface FetchAPI
 */
export interface FetchAPI {
    (url: string, init?: any): Promise<Response>;
}

/**
 *
 * @export
 * @interface FetchArgs
 */
export interface FetchArgs {
    url: string;
    options: any;
}

/**
 *
 * @export
 * @class BaseAPI
 */
export class BaseAPI {
    protected configuration: Configuration | undefined;

    constructor(configuration?: Configuration, protected basePath: string = BASE_PATH, protected fetch: FetchAPI = isomorphicFetch) {
        if (configuration) {
            this.configuration = configuration;
            this.basePath = configuration.basePath || this.basePath;
        }
    }
}
// ...
export interface LoginRequest {
    /**
     * 
     * @type {string}
     * @memberof LoginRequest
     */
    username?: string;
    /**
     * 
     * @type {string}
     * @memberof LoginRequest
     */
    password?: string;
}
// ...
export class EstablishmentsApi extends BaseAPI {
    /**
     * 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof EstablishmentsApi
     */
    public apiEstablishmentsGet(options?: any) {
        return EstablishmentsApiFp(this.configuration).apiEstablishmentsGet(options)(this.fetch, this.basePath);
    }

    /**
     * 
     * @param {string} id 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof EstablishmentsApi
     */
    public apiEstablishmentsIdDelete(id: string, options?: any) {
        return EstablishmentsApiFp(this.configuration).apiEstablishmentsIdDelete(id, options)(this.fetch, this.basePath);
    }

    /**
     * 
     * @param {string} id 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof EstablishmentsApi
     */
    public apiEstablishmentsIdGet(id: string, options?: any) {
        return EstablishmentsApiFp(this.configuration).apiEstablishmentsIdGet(id, options)(this.fetch, this.basePath);
    }

    /**
     * 
     * @param {string} id 
     * @param {Establishment} [body] 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof EstablishmentsApi
     */
    public apiEstablishmentsIdPut(id: string, body?: Establishment, options?: any) {
        return EstablishmentsApiFp(this.configuration).apiEstablishmentsIdPut(id, body, options)(this.fetch, this.basePath);
    }

    /**
     * 
     * @param {string} name 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof EstablishmentsApi
     */
    public apiEstablishmentsNameNameGet(name: string, options?: any) {
        return EstablishmentsApiFp(this.configuration).apiEstablishmentsNameNameGet(name, options)(this.fetch, this.basePath);
    }

    /**
     * 
     * @param {boolean} open 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof EstablishmentsApi
     */
    public apiEstablishmentsOpenOpenGet(open: boolean, options?: any) {
        return EstablishmentsApiFp(this.configuration).apiEstablishmentsOpenOpenGet(open, options)(this.fetch, this.basePath);
    }

    /**
     * 
     * @param {Establishment} [body] 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof EstablishmentsApi
     */
    public apiEstablishmentsPost(body?: Establishment, options?: any) {
        return EstablishmentsApiFp(this.configuration).apiEstablishmentsPost(body, options)(this.fetch, this.basePath);
    }

    /**
     * 
     * @param {number} [minScore] 
     * @param {number} [maxScore] 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof EstablishmentsApi
     */
    public apiEstablishmentsScoreRangeGet(minScore?: number, maxScore?: number, options?: any) {
        return EstablishmentsApiFp(this.configuration).apiEstablishmentsScoreRangeGet(minScore, maxScore, options)(this.fetch, this.basePath);
    }

    /**
     * 
     * @param {number} score 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof EstablishmentsApi
     */
    public apiEstablishmentsScoreScoreGet(score: number, options?: any) {
        return EstablishmentsApiFp(this.configuration).apiEstablishmentsScoreScoreGet(score, options)(this.fetch, this.basePath);
    }

}
// ...
export class LoginApi extends BaseAPI {
    /**
     * 
     * @param {LoginRequest} [body] 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof LoginApi
     */
    public apiLoginPost(body?: LoginRequest, options?: any) {
        return LoginApiFp(this.configuration).apiLoginPost(body, options)(this.fetch, this.basePath);
    }

}
// ...
export class ProductsApi extends BaseAPI {
    /**
     * 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof ProductsApi
     */
    public apiProductsGet(options?: any) {
        return ProductsApiFp(this.configuration).apiProductsGet(options)(this.fetch, this.basePath);
    }

    /**
     * 
     * @param {string} id 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof ProductsApi
     */
    public apiProductsIdDelete(id: string, options?: any) {
        return ProductsApiFp(this.configuration).apiProductsIdDelete(id, options)(this.fetch, this.basePath);
    }

    /**
     * 
     * @param {string} id 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof ProductsApi
     */
    public apiProductsIdGet(id: string, options?: any) {
        return ProductsApiFp(this.configuration).apiProductsIdGet(id, options)(this.fetch, this.basePath);
    }

    /**
     * 
     * @param {string} id 
     * @param {Product} [body] 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof ProductsApi
     */
    public apiProductsIdPut(id: string, body?: Product, options?: any) {
        return ProductsApiFp(this.configuration).apiProductsIdPut(id, body, options)(this.fetch, this.basePath);
    }

    /**
     * 
     * @param {Product} [body] 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof ProductsApi
     */
    public apiProductsPost(body?: Product, options?: any) {
        return ProductsApiFp(this.configuration).apiProductsPost(body, options)(this.fetch, this.basePath);
    }

    /**
     * 
     * @param {string} id 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof ProductsApi
     */
    public apiProductsPublicationIdGet(id: string, options?: any) {
        return ProductsApiFp(this.configuration).apiProductsPublicationIdGet(id, options)(this.fetch, this.basePath);
    }

}
// ...
export class PublicationsApi extends BaseAPI {
    /**
     * 
     * @param {string} id 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof PublicationsApi
     */
    public apiPublicationsEstablishmentIdGet(id: string, options?: any) {
        return PublicationsApiFp(this.configuration).apiPublicationsEstablishmentIdGet(id, options)(this.fetch, this.basePath);
    }

    /**
     * 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof PublicationsApi
     */
    public apiPublicationsGet(options?: any) {
        return PublicationsApiFp(this.configuration).apiPublicationsGet(options)(this.fetch, this.basePath);
    }

    /**
     * 
     * @param {string} id 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof PublicationsApi
     */
    public apiPublicationsIdDelete(id: string, options?: any) {
        return PublicationsApiFp(this.configuration).apiPublicationsIdDelete(id, options)(this.fetch, this.basePath);
    }

    /**
     * 
     * @param {string} id 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof PublicationsApi
     */
    public apiPublicationsIdGet(id: string, options?: any) {
        return PublicationsApiFp(this.configuration).apiPublicationsIdGet(id, options)(this.fetch, this.basePath);
    }

    /**
     * 
     * @param {string} id 
     * @param {Publication} [body] 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof PublicationsApi
     */
    public apiPublicationsIdPut(id: string, body?: Publication, options?: any) {
        return PublicationsApiFp(this.configuration).apiPublicationsIdPut(id, body, options)(this.fetch, this.basePath);
    }

    /**
     * 
     * @param {Publication} [body] 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof PublicationsApi
     */
    public apiPublicationsPost(body?: Publication, options?: any) {
        return PublicationsApiFp(this.configuration).apiPublicationsPost(body, options)(this.fetch, this.basePath);
    }

    /**
     * 
     * @param {number} [minPrice] 
     * @param {number} [maxPrice] 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof PublicationsApi
     */
    public apiPublicationsPriceGet(minPrice?: number, maxPrice?: number, options?: any) {
        return PublicationsApiFp(this.configuration).apiPublicationsPriceGet(minPrice, maxPrice, options)(this.fetch, this.basePath);
    }

    /**
     * 
     * @param {number} [minScore] 
     * @param {number} [maxScore] 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof PublicationsApi
     */
    public apiPublicationsScoreGet(minScore?: number, maxScore?: number, options?: any) {
        return PublicationsApiFp(this.configuration).apiPublicationsScoreGet(minScore, maxScore, options)(this.fetch, this.basePath);
    }

    /**
     * 
     * @param {string} id 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof PublicationsApi
     */
    public apiPublicationsUserIdGet(id: string, options?: any) {
        return PublicationsApiFp(this.configuration).apiPublicationsUserIdGet(id, options)(this.fetch, this.basePath);
    }

}
// ...
export class UsersApi extends BaseAPI {
    /**
     * 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof UsersApi
     */
    public apiUsersGet(options?: any) {
        return UsersApiFp(this.configuration).apiUsersGet(options)(this.fetch, this.basePath);
    }

    /**
     * 
     * @param {string} id 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof UsersApi
     */
    public apiUsersIdDelete(id: string, options?: any) {
        return UsersApiFp(this.configuration).apiUsersIdDelete(id, options)(this.fetch, this.basePath);
    }

    /**
     * 
     * @param {string} id 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof UsersApi
     */
    public apiUsersIdGet(id: string, options?: any) {
        return UsersApiFp(this.configuration).apiUsersIdGet(id, options)(this.fetch, this.basePath);
    }

    /**
     * 
     * @param {string} id 
     * @param {User} [body] 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof UsersApi
     */
    public apiUsersIdPut(id: string, body?: User, options?: any) {
        return UsersApiFp(this.configuration).apiUsersIdPut(id, body, options)(this.fetch, this.basePath);
    }

    /**
     * 
     * @param {User} [body] 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof UsersApi
     */
    public apiUsersPost(body?: User, options?: any) {
        return UsersApiFp(this.configuration).apiUsersPost(body, options)(this.fetch, this.basePath);
    }

}